Example Log Return Time Series

A tibble containing daily negative log returns of closing prices for the S&P 500 stock market index. The observation period spans 20 trading years: 1995-01-01 to 2024-12-31.

Usage

data(logret_data)

Format

A tibble with 7,550 rows and 2 columns:

day

Date of observation (class Date)

neg_log_ret

Negative log return (numeric)

Details

The data was obtained using the quantmod package with Yahoo Finance as the source.